test: move DefaultHttpDecodeResponseTest to common>encoding path and change the name of test classes
